Rename folder containing Test Explorer items more appropriately.
