+++ src/client/unittests/explorer/commandHandlers.ts
